Extracting the spec helper to its own class in lib so we can leverage it in chef-provisioning for testing there
