+++ cheffish.gemspec
@@ -12,7 +12,8 @@ Gem::Specification.new do |s|
   s.email = 'jkeiser@opscode.com'
   s.homepage = 'http://wiki.opscode.com/display/chef'
 
-  s.add_dependency 'chef-zero'
+  s.add_dependency 'chef-zero', '~> 4.0'
+  s.add_dependency 'chef' #, '~> 12.1'
 
   s.add_development_dependency 'rake'
   s.add_development_dependency 'rspec', '~> 3.0'
