+++ spec/integration/chef_client_spec.rb
@@ -1,4 +1,5 @@
 require 'support/spec_support'
+require 'chef/support/chef_run_support'
 require 'support/key_support'
 require 'chef/resource/chef_client'
 require 'chef/provider/chef_client'
@@ -6,7 +7,7 @@ require 'chef/provider/chef_client'
 repo_path = Dir.mktmpdir('chef_repo')
 
 describe Chef::Resource::ChefClient do
-  extend SpecSupport
+  extend Chef::Support::ChefRunSupport
 
   when_the_chef_12_server 'is in multi-org mode' do
     organization 'foo'
