+++ spec/unit/get_private_key_spec.rb
@@ -1,4 +1,5 @@
 require 'support/spec_support'
+require 'chef/support/chef_run_support'
 
 describe Cheffish do
   let(:directory_that_exists) {
